Fixed problem with placing tile render order.
Before, it would immediately drop below the score layer and fly to its position
on the board. Now we have a "flight" layer which renders above the score layer,
where it does its flying into place, and only when it arrives is it moved to
the tiles layer. The flight layer is kept at the same translation as the tiles
layer so that funny business does not happen if the tiles layer is scrolled
while the tile is flying into place.