When not animating, set trans/rot immediately on glyphs.
This was causing problems with other animations because piecen scoring triggers
would show up and the just placed glyph would not be properly rotated or
translated because the 0 second animation was delayed by a barrier. So now when
we want to apply our bits immediately, we do so directly rather than via a 0
second animation. That way the changes are not blocked by a barrier and when
the scoring animation preparations take place, they use the correct values.