Clean up local project handling.
I tried valiantly to allow local projects to be magically inherited, but due to
limitations in my understandings of the internals of SBT, I didn't get the
whole job done. Presently one has to create symlinks for transitive local
projects in the current project's directory (though they need not be declared
in the current project's locals). So atlantis inherits forplay from tripleplay,
but it needs a forplay symlink in its top-level directory for everything to
work. Similarly (in this case) for react and pythagoras.