Nix GWT bits in SBT build. Update Guava version.
Having the GWT stuff in the SBT build is an ongoing PITA (mostly thanks to the
frequently changing plugins and SBT's unfortunate requirement that its plugins
be dependent on a three way matrix of Scala version, SBT version, and plugin
version).

One can build the HTML backend via Maven easily enough. SBT's main value add is
for rapid testing with the Java backend with just-in-time automatic
recompilation of dependent projects.